dataload

+++ app/src/main/java/com/mobilpogbead/database/Score.kt
@@ -1,3 +1,3 @@
 package com.mobilpogbead.database
 
-data class Score(val name:String,val score:Int,val time:Long) {}
+data class Score(val name:String,val score:Int,val time:Float) {}
